fix project module organisation and type comments
