--- src/main/python/covata/delta/api/api_client.py
@@ -1,51 +1,0 @@
-#   Copyright 2016 Covata Limited or its affiliates
-#
-#   Licensed under the Apache License, Version 2.0 (the "License");
-#   you may not use this file except in compliance with the License.
-#   You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-
-from abc import ABCMeta, abstractmethod
-
-
-class ApiClient(object):
-    __metaclass__ = ABCMeta
-
-    DELTA_URL = 'https://delta.covata.cc/master'
-    RESOURCE_IDENTITIES = '/identities'
-
-    def __init__(self, crypto_service):
-        """
-        Constructs a new Delta API client with the given configuration
-
-
-        :param crypto_service: the CryptoService object
-        :type crypto_service: :class:`~covata.delta.crypto.CryptoService`
-        """
-        self.crypto_service = crypto_service
-
-    @abstractmethod
-    def register_identity(self, external_id=None, metadata=None):
-        # (ApiClient, Optional[str], Optional[dict]) -> str
-        """
-        Creates a new identity in Delta with the provided metadata
-        and external id.
-
-        :param optional(str) external_id:
-            the external id to associate with the identity
-
-        :param optional(dict) metadata:
-            the metadata to associate with the identity
-
-        :return: the id of the newly created identity
-
-        :rtype: str
-        """
-        pass
