+++ src/main/python/covata/delta/crypto.py
@@ -22,10 +22,10 @@ from cryptography.hazmat.primitives.asymmetric import rsa
 
 
 class CryptoService:
-
     def __init__(self,
                  key_store_path,
                  key_store_passphrase):
+        # type: (str, str) -> self
         """
         Constructs a new Crypto service with the given configuration.
 
@@ -35,23 +35,27 @@ class CryptoService:
         self.key_store_path = os.path.expanduser(key_store_path)
         self.key_store_passphrase = key_store_passphrase
 
-    def write_to_file(self, private_key, file_name):
-        # (Crypto, RSAPrivateKey, str) -> None
+    def save(self, private_key, file_name):
+        # type: (rsa.RSAPrivateKey, str) -> None
         pem = private_key.private_bytes(
             serialization.Encoding.PEM,
             serialization.PrivateFormat.PKCS8,
             serialization.BestAvailableEncryption(self.key_store_passphrase))
 
         file_path = os.path.join(self.key_store_path, file_name)
-
         if not os.path.isdir(self.key_store_path):
             os.makedirs(self.key_store_path)
 
         with open(file_path, 'w') as f:
             f.write(pem)
 
-    def read_from_file(self, file_name):
-        # (Crypto, str) -> RSAPrivateKey
+    def load(self, file_name):
+        # type: (str) -> rsa.RSAPrivateKey
+        """
+
+        :param file_name:
+        :return:
+        """
         file_path = os.path.join(self.key_store_path, file_name)
         with(open(file_path, 'r')) as f:
             return serialization \
@@ -61,7 +65,7 @@ class CryptoService:
 
     @staticmethod
     def generate_key():
-        # () -> RSAPrivateKey
+        # type: () -> rsa.RSAPrivateKey
         """
         Generates an RSA private key object. The public key object can be
         extracted by calling public_key() method on the generated key object.
@@ -77,7 +81,7 @@ class CryptoService:
 
     @staticmethod
     def serialized(public_key):
-        # (RSAPublicKey) -> String
+        # type: (rsa.RSAPublicKey) -> str
         """
 
         :param :class:`RSAPublicKey` public_key: the public Key object
